Redactored code to have clean code and updated readme
