+++ index.html
@@ -14,24 +14,26 @@
   </head>
   <body>
     <nav class="nav_bar">
-      <div class="nav_logo">
-        <a href="/">
-          <img src="assets/icons/logo.png" alt="logo" class="logo" />
-        </a>
+      <div class="nav_header">
+        <div class="nav_logo">
+          <a href="/">
+            <img src="assets/icons/logo.png" alt="logo" class="logo" />
+          </a>
+        </div>
+        <ul class="nav_links" id="nav_links">
+          <li><a href="pages/log-in.html">Login</a></li>
+          <li><a href="pages/sign-up.html">Let's Chat</a></li>
+        </ul>
+        <button
+          class="menu_btn"
+          id="menu_btn"
+          aria-label="Toggle navigation menu"
+          aria-expanded="false"
+        >
+          <i class="ri-menu-line"></i>
+        </button>
       </div>
-          <ul class="nav_links" id="nav_links">
-            <li><a href="pages/log-in.html">Login</a></li>
-            <li><a href="pages/sign-up.html">Let's Chat</a></li>
-          </ul>
-          <button
-            class="menu_btn"
-            id="menu_btn"
-            aria-label="Toggle navigation menu"
-            aria-expanded="false"
-          >
-            <i class="ri-menu-line"></i>
-          </button>
-      </nav>
+    </nav>
     <main>
       <section class="section_container header_container">
         <div class="header_content">
