+++ js/utils/Logger.js
@@ -1,0 +1,21 @@
+
+
+class Logger {
+    error(message, error = null) {
+        if (error) {
+            console.error(`[ERROR] ${message}`, error);
+        } else {
+            console.error(`[ERROR] ${message}`);
+        }
+    }
+
+    warn(message, data = null) {
+        if (data) {
+            console.warn(`[WARN] ${message}`, data);
+        } else {
+            console.warn(`[WARN] ${message}`);
+        }
+    }
+}
+
+export const logger = new Logger();
