+++ pages/chat.html
@@ -8,6 +8,7 @@
     <link rel="stylesheet" href="../styles/style.css">
     <link rel="stylesheet" href="../styles/chat.css">
     <link rel="stylesheet" href="../styles/group-chat.css">
+    <link rel="stylesheet" href="../styles/notification.css">
     <link
   href="https://cdn.jsdelivr.net/npm/remixicon@4.9.0/fonts/remixicon.css"
   rel="stylesheet"
