+++ pages/forgot-password.html
@@ -7,6 +7,7 @@
     <title>ChatX - Forgot Password</title>
     <link rel="stylesheet" href="../styles/style.css">
     <link rel="stylesheet" href="../styles/auth.css">
+    <link rel="stylesheet" href="../styles/notification.css">
     <style>
         .hidden { display: none; }
     </style>
